write getCurrentCategories service function
