+++ services/itemService.js
@@ -6,13 +6,11 @@ export async function getAllItems() {
 
   const items = await Item.find();
 
-  const sanitizedItems = items.map(({ id, name, category }) => {
-    return {
-      id: id,
-      name: name,
-      category: { name: category.name, icon_src: category.icon_src },
-    };
-  });
+  const sanitizedItems = items.map(({ id, name, category }) => ({
+    id: id,
+    name: name,
+    category: { name: category.name, icon_src: category.icon_src },
+  }));
 
   return sanitizedItems;
 }
